Removed webpack dev server, moved backend server and set it to serve React views
